Move Color to its own file
